requirements.txt
# Dynamic Skills System Dependencies
# Version: 1.0.0
# Python: 3.10+

# Core dependencies
psycopg2-binary>=2.9.9
openai>=1.12.0
redis>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Vector search
pgvector>=0.2.4

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1

# Web service
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Optional: For skills scanner
beautifulsoup4>=4.12.0
lxml>=5.1.0
